requirements.txt
torch>=2.3.0
transformers>=4.45.0
trl>=0.12.0
peft>=0.12.0
datasets>=2.20.0
accelerate>=0.34.0
bitsandbytes>=0.43.0
sympy>=1.13.0
vllm>=0.6.0
wandb>=0.18.0
numpy>=1.26.0
tqdm>=4.66.0
pyyaml>=6.0